pyproject.toml
[tool.poetry]
name = "ihrisplus"
version = "0.1.0"
description = "A Flask + Angular dashboard application"
authors = ["cheikhou <niangcheikhou@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.0"
flask-restplus = "^0.12.1"
sqlalchemy = "^1.2"
python-dotenv = "^0.10.0"
MySQL-connector-python = "^8.0"
pandas = "^0.23.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
